+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/openid/RedirectUriClientTest.kt
@@ -292,7 +292,7 @@ val RedirectUriClientTest by matrixSuite {
         }
 
         "test with deserializing" {
-            val authnRequest = it.verifierOid4vp.createAuthnRequest(defaultRequestOptions)
+            val authnRequest = it.verifierOid4vp.createPlainAuthnRequest(defaultRequestOptions)
             val authnRequestUrlParams = authnRequest.encodeToParameters().formUrlEncode()
 
             val parsedAuthnRequest: AuthenticationRequestParameters =